Generally make VirtualGL more friendly when it is installed from source.  vglrun will now load the faker libraries from whatever directory they reside in, assuming that the directory was specified in the VGL_LIBDIR CMake variable.  This is accomplished via two new scripts, .vglrun.vars32 for a 32-bit build and .vglrun.vars64 for a 64-bit build.  vglrun will execute either one if it exists, which allows it to add the value of VGL_LIBDIR from both a 32-bit and 64-bit install to LD_LIBRARY_PATH.  Since the official packages place the faker libraries in system library directories (necessary so setuid root executables can be launched in VirtualGL), the new scripts are not necessary and thus aren't included in those packages.
If the default install prefix (/opt/VirtualGL) is used, then the build system now sets other defaults to enable a 32-bit and a 64-bit install of VirtualGL to be co-located.  Specifically, the default value of VGL_LIBDIR and VGL_FAKELIBDIR are set to /opt/VirtualGL/lib32 and /opt/VirtualGL/fakelib32 for 32-bit builds, and the default value of VGL_LIBDIR and VGL_FAKELIBDIR are set to /opt/VirtualGL/lib64 and /opt/VirtualGL/fakelib64 for 64-bit builds.  Also, glxspheres is renamed to glxspheres64 when installing a 64-bit build.


git-svn-id: svn+ssh://svn.code.sf.net/p/virtualgl/code/trunk@3387 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b