osacompile generates the Uninstall VirtualGL app as a universal binary by default.  However, perhaps because we're building it on a newer version of OS X, older OS X versions (in particular, Leopard) don't like our universal binary and report it as being of an "unsupported architecture."  The easiest way around this is to just force osacompile to build an i386 binary.
Remove sudo and chown commands (no longer needed with pkgbuild/productbuild.)

Create a link to the User's Guide in the /Applications/VirtualGL folder, rather than a link to the entire documentation directory.  Place a copy of the Unininstall app in /Applications/VirtualGL as well.  Both of these changes mimic the behavior of the TurboVNC package.


git-svn-id: svn+ssh://svn.code.sf.net/p/virtualgl/code/branches/2.3.x@3437 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b