Modern versions of Xlib implement XNextEvent() by locking the XCB connection and calling xcb_wait_for_event().  Linux implementations of Qt5 are built upon glib, so when Qt5 is processing events, it will also call down to the glib event handler, which calls XNextEvent().  Thus, when VirtualGL attempts to call xcb_intern_atom_reply() within the body of xcb_wait_for_event(), it deadlocks because the connection is already locked by Xlib.  The best approach for now is simply to cache the value of the atoms in the body of XOpenDisplay() instead.
Note that this highlights a concern I had regarding interposing XCB-- since Xlib is implemented on top of XCB, there are situations in which VirtualGL may actually be interposing twice, and this appears to be one of them.  If the application is calling XPending() or another Xlib function to query events, then VirtualGL need not handle the events at the XCB level in that case.  For now, the duplication of effort does not seem to cause problems, but it may be necessary to do something about it in the future.


git-svn-id: svn+ssh://svn.code.sf.net/p/virtualgl/code/trunk@3590 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b