Server: fix error when running Firefox
Firefox (tested with v31 and later) would generate the following error
when running in VirtualGL:

[VGL] ERROR: in readPixels--
[VGL]    XXX: VirtualDrawable instance has not been fully initialized

The application apparently calls XCopyArea() and passes a GLX Pixmap
handle as the source drawable, but it does this before the source
drawable has been made current.  Thus, from VirtualGL's point of view,
the virtual drawable associated with the GLX Pixmap has not yet been
initialized.  The solution was to ignore the VirtualDrawable associated
with the source or destination drawable passed to XCopyArea() unless
that VirtualDrawable has been initialized.  If it hasn't, then we can
safely assume that no 3D pixels have been drawn to the drawable, so we
can just use the plain X11 XCopyArea() function.