Faker: Fix thread safety issue in PBO readback
In C++, static variables (even ones within the scope of a method) have
the same value across all instances, so in fact VirtualGL was attempting
to use the same PBO for every drawable.  This was suspected to cause an
intermittent error in MetaPost:

[VGL] ERROR: in readPixels--
[VGL]    Could not set PBO size

when the application started, probably due to a race condition.