Faker: Temp. workaround for FF OMTC BadMatch error
Version 325.08 of the nVidia drivers began generating BadMatch errors
when glXCreatePixmap() was called with a Pixmap whose depth did not
match the depth of the GLXFBConfig passed to the same function.  This
most notably caused recent versions of Firefox with off-main-thread
compositing to fail, because Firefox creates a depth-32 Pixmap but then
passes the first GLXFBConfig returned from glXChooseFBConfig() (which is
a depth-24 GLXFBConfig) into glXCreatePixmap().

Referring to
https://bugzilla.mozilla.org/show_bug.cgi?id=896287

Mozilla has long since worked around the issue by detecting the presence
of the nVidia driver and changing the Firefox visual selection behavior
accordingly, but VirtualGL breaks this detection mechanism by setting
the GLX vendor strings to "VirtualGL".  And for reasons that are as yet
unknown, setting the GLX vendor string to "NVIDIA Corporation" causes
Firefox to break in even worse ways.  This patch at least allows it to
work, but until we can figure out the underlying causes of these errors,
I'm not comfortable putting this into production yet.

Works around #3