Faker: Allow GLX vendor string to be specified
Provide an undocumented environment variable (VGL_GLXVENDOR) that causes
VirtualGL to return a specified string when the application calls
glXGetClientString(..., GLX_VENDOR) or
glXQueryServerString(..., GLX_VENDOR).  At the moment, the only known
use for this is testing the Firefox OMTC bug (#3).  Firefox looks at the
GLX vendor string to determine how to select visuals, and if it sees
"NVIDIA Corporation", it will choose a 32-bit GLXFBConfig to match the
depth-32 Pixmap.  Otherwise, it will pass a 24-bit GLXFBConfig with a
depth-32 Pixmap to glXCreatePixmap(), which results in a BadMatch error
with nVidia 325.08 and later (but not other drivers.)

Setting VGL_GLXVENDOR="NVIDIA Corporation" works around the BadMatch
error, but unfortunately this is insufficient to fix the OMTC rendering
issue.