Faker: Ensure globalMutex initialized prior to use
In some cases (known to be the case with MainWin, for instance), a
shared library's constructor (_init() or a function with the GCC
"constructor" attribute) can call one of the interposed functions
(XOpenDisplay(), for instance) before VirtualGL's own _init() ->
__static_initialization_and_destruction_0() is called.  Thus,
vglfaker::init() will be called prior to the globalMutex constructor,
and the pthread mutex within globalMutex will still be a fast
(PTHREAD_MUTEX_NORMAL) mutex instead of a recursive mutex, resulting in
a deadlock.

This is known to affect MATLAB's need_softwareopengl utility and Ansys
DesignModeler 16.1+.

To work around this, a singleton CriticalSection instance was created
for globalMutex, and this instance is initialized on first use.

Adapted from nathankidd/virtualgl@13fa39e3a
Closes #7