Faker: Don't instantiate GCS in global destructor
For unexplained reasons, on certain platforms (referring to #12, known
to be the case with recent spins of Arch Linux-- glibc 2.22, GCC 5.2),
pthread mutexes cease to work properly by the time the faker's global
destructor (GlobalCleanup::~GlobalCleanup()) is called.  Because of the
MainWin/GlobalCriticalSection fix, globalMutex is instantiated on first
use, and if a particular application isn't taking advantage of any
functions that VirtualGL interposes, that can lead to a situation in
which the faker tries to instantiate globalMutex in the body of
GLobalCleanup::~GlobalCleanup().  In this case, the SafeLock within
GlobalCriticalSection::getInstance() will fail, causing the application
to crash due to an unhandled exception.

There are a couple of ways that this could be worked around, but the
least intrusive seems to be to check whether GlobalCriticalSection has
been instantiated.  If not, then there is no reason to lock it in the
global destructor.