Faker: Fix regression (segfault) on Solaris
glXGetProcAddress() on Solaris will, for unknown reasons,
return the address of VirtualGL's interposed version of a GLX/OpenGL
function if the underlying "real" version of that function doesn't
exist in libGL.  If this happens, then the interposed function will
keep recursing until the program blows its stack and segfaults.  Thus,
we use the old (VGL <= 2.4.x) method of loading the "real" OpenGL
functions on Solaris.  This should have no negative ramifications unless
a program is particularly ill-behaved (SPECviewperf, for instance) and
tries to override one of the OpenGL/GLX symbols at file scope.

NOTE: I can only test Solaris in a virtual machine, so the underlying
OpenGL library is Mesa, but I verified that this issue doesn't occur
on a simiarly-configured Linux VM.  It seems to be a Solaris-specific
issue (perhaps due to differing dlsym() behavior.)  It is unknown
whether the issue also exists on Solaris with the nVidia drivers (which,
BTW, are the only reason why we still support Solaris.)