Faker: Don't report exts. that libGL doesn't have
VirtualGL 2.4 (specifically, edc59e3de2d42a29693e09bf099a9359d148d8bc)
introduced code into glXGetProcAddress() that, for optional GLX
extensions, checks for the presence of the underlying function symbols
in libGL before returning the corresponding interposed symbols.
However, for certain extensions, VirtualGL was always reporting the
presence of the extension to the 3D application, regardless of whether
libGL contained the necessary functions to support the extension.  This
created a regression with WINE, since it apparently probes for the
presence of GLX_*_swap_interval and, if available, blindly invokes the
function pointer returned from glXGetProcAddress("glXSwapIntervalEXT")
or glXGetProcAddress("glXSwapIntervalSGI") (which is NULL if libGL
doesn't have those functions.)

This commit modifies the faker such that it only reports an optional GLX
extension to the 3D application if libGL contains the necessary
functions to support the extension.

Fixes #21