Faker: Connect to 3D X server only when needed
This reduces the overhead of the faker when running non-OpenGL
applications.  Note also that this commit removes the crufty is3D()
macro that was previously used to guard against recursion.  Instead,
the GLX and X11 interposers now use the same technique as the XCB
interposer to guard against recursion (the faker level TLS variable.)
This should be more robust, and it has the added benefit of not
requiring the 3D X server connection to be open.