Faker: Correct type mismatch of deadYet variable
deadYet was made a bool in 7af4f87a066135aa4acff734a494dce3fbebafba,
and the extern reference to it should have been changed to a bool as
well in a7a910167b6c4b99252c3308aaa706f115115cb0.  This bug escaped
detection until recently, but for reasons not yet fully understood, the
32-bit optimizer in recent versions of GCC caused the bug to surface in
unpredictable ways.  Under certain circumstances, the extern int value
of deadYet would take on a value other than 0 or 1, which caused a false
positive in sections of the code that check this value.  More
specifically, I observed that the changes made in
262e077ed9f21ebb3176e556b32821c7f8400c9c (on-demand connections to the
3D X server) caused this problem to surface with a 32-bit build of
VirtualGL built with GCC 6.2.  The result was that VirtualGL's
interposed version of XQueryExtension() sometimes erroneously passed
through to the "real" XQueryExtension() function, and this caused libGL
(nVidia) to believe that GLX was not supported by the X server
(TurboVNC, which lacks a GLX extension.)  Others have reported that this
same bug caused Steam to crash, although I wasn't able to reproduce
that.

Long and the short of it-- the mismatch between deadYet types across
different execution units led to unpredictable faker behavior in some
cases.

Fixes #31, Closes #33