Build: Fix VGL_BUILDSTATIC w/ FreeBSD & Clang
VGL_BUILDSTATIC was disabled by default on FreeBSD because it
apparently didn't work with FreeBSD 8.x, but it seems to work fine with
currently supported versions (I test with 10.3 at the moment.)

This also enables VGL_BUILDSTATIC for Clang builds.  On Linux at least,
it works fine as long as we pass -static-libgcc only to the linker
(tested with Clang 3.8 on Fedora.)  Doesn't seem to work with Clang 3.4
on FreeBSD, but VGL still builds properly otherwise.