Faker: Always skip readback if mode != GL_RENDER
The previous commit introduced a conformance issue, because
glRenderMode(GL_RENDER) resets the state of the feedback or select
buffer.  Referring to the discussion in #41, it would be possible to
interpose glRenderMode(), thus allowing us to internally capture the
number of hit records or values transferred to the select/feedback
buffer and pass back that value to the 3D application whenever it calls
glRenderMode(GL_RENDER).  However, calling glRenderMode(GL_RENDER)
internally to VGL (within the body of glXSwapBuffers()) also resets the
select/feedback buffer pointer, and there is no way to restore that
pointer to its previous state.  Under the following scenario, for
instance:

    GLfloat values[6] = { 0, 0, 0, 0, 0, 0 };
    glFeedbackBuffer(6, GL_3D, values);
    glRenderMode(GL_FEEDBACK);
    glPassThrough(1.0);
    glPassThrough(2.0);
    glXSwapBuffers(dpy, drawable);
    glPassThrough(3.0);
    numTransferred=glRenderMode(GL_RENDER);

the previous commit would return an incorrect value for numTransferred.
By interposing glRenderMode(), VirtualGL could return the correct value
(6) for numTransferred, but the modifications would cause the values
array to be { 1792.0, 3.0, 1792.0, 2.0, 0.0, 0.0 } instead of
{ 1792.0, 1.0, 1792.0, 2.0, 1792.0, 3.0 }.  The only methods we could
devise to work around this were:

1. Use a separate thread for readback, thus avoiding the need to switch
contexts within the rendering thread.

2. Only perform the readback under certain limited conditions if
render mode != GL_RENDER (the readback and render drawables must
be identical, and the OpenGL state must be conducive to performing
the readback.)

3. Always skip the readback if render mode != GL_RENDER.

(1) is the most conformant option but also the most difficult to
implement, and the most likely to introduce bugs.  (3) is the easiest
and least disruptive option.  Although it is not technically conformant,
it causes glXSwapBuffers() to behave as glFinish(), glFlush(), and
glXWaitGL() already behave, and this seems to cause no problems with
the application that originally inspired this fix.  If this behavior
proves to cause problems with other applications, then we can look into
implementing (1) down the road.

Closes #41