Build: Don't expose VGL_BUILDSERVER on Mac, Cygwin
In the process of cleaning up the build system in
c35b13d0530f8728f2440d88149bdee15648cbf2, VGL_BUILDSERVER was exposed on
all Un*x platforms.  This commit reverts to the previous behavior (only
exposing that variable on non-Apple and non-Cygwin platforms.)  The VGL
server components do not even pretend to work on OS X or Cygwin, nor is
there any reason to build those components on those platforms.