Faker: Don't handle WM_DELETE_WINDOW if app does
Because VGL is interposing XNextEvent() and friends, it is notified of
WM_DELETE_WINDOW messages before the application is, so even if the
application was handling WM_DELETE_WINDOW messages, VirtualGL would
still throw an error if the application didn't immediately shut down its
X11 event queue in response to one of those messages.  This commit
causes VGL to ignore WM_DELETE_WINDOW messages altogether if it detects
that the application is already handling them.

VirtualGL's handling of those messages was meant to prevent BadDrawable
errors that would occur in the X11 Transport if the WM closed the window
without warning, but the WM will never close the window if the 3D
application is handling WM_DELETE_WINDOW messages.

Fixes #47