Ensure that Frame.pf is always initialized
Oversight from 7b802df6f1a7b7623598d90a2f69c7de580b8402.  It's easier to
just initialize the pointer to &__format_NONE than to check whether it's
NULL in all of the places that refer to it.  Oddly, this issue didn't
show up except when running vglclient on a Mac, and it manifested as an
"Invalid argument" error in Event::wait().  valgrind on Linux didn't
catch it.