Work around PV segfault w/ non-GLVND nVidia driver
The proximate cause of this is the fact that ParaView calls
glXDestroyContext() to destroy a rendering context in which the front
buffer is the active draw buffer, then it calls
glXMakeCurrent(..., 0, 0) to swap out the context.  When the application
calls glXMakeCurrent(..., 0, 0), VirtualGL detects that front buffer
rendering has occurred but the front buffer has not been read back yet,
so VGL tries to read back and transport the contents of the buffer.  To
do this, VirtualGL swaps in a temporary "readback context", but when VGL
tries to swap back in the previous context (which had been destroyed but
was still current prior to the readback), a segfault occurs within the
bowels of nVidia's libGL.  I would argue that this segfault is a driver
bug and that nVidia's GLX implementation should instead throw a
GLXBadContext error.  The segfault goes away when using the
GLVND-enabled version of the driver, which lends further credence to the
argument that it is a driver bug.

Furthermore, this segfault is reproducible with a simple test program,
irrespective of VirtualGL (refer to #44.)  However, I also admit that
VirtualGL shouldn't be trying to call glXMakeContextCurrent() on a
context that has already been destroyed, so this commit prevents VGL
from doing that, thus working around the segfault.

Fixes #44