Doc: Clarify the GDM 3 Init/Default problem
Referring to the conversation in
https://gitlab.gnome.org/GNOME/gdm/issues/317
this behavior was a consequence of migrating to Wayland, and for a
variety of reasons, it is unlikely to change.

NOTE: Referring to the same conversation, it seems that the
display-setup-script problem with LightDM in Ubuntu 18.04 was also
expected behavior, a consequence of using a separate X server instance
for the login prompt and the user session.