fakerut: Improve stability of multithreaded test
getenv() is only thread-safe if setenv()/putenv() is not called in a
different thread, and since that is exactly what the Faker autotest
mechanism was doing, the multithreaded test frequently segfaulted.  This
commit modifies the autotest mechanism so that VirtualDrawable stores
the window color and frame count in thread-local variables and frameut
obtains those values from dynamically-loaded functions.