Faker: Don't call glXChooseFBConfig() w/ _dpy3D
We did this in matchConfig() solely for the purpose of tracing the
internal glXChooseFBConfig() call, which could provide useful diagnostic
information if a particular "visual hunting" application failed to
obtain a visual.  However, this also caused VirtualGL to create an
unnecessary visual attribute table for the 3D X server, among other
things.  This commit modifies matchConfig() so that it calls the
non-interposed version of glXChooseConfig() and outputs the necessary
tracing information itself.