Remove support for transparent overlays
This is necessary in order to make way for upcoming modifications to
VirtualGL to support X-server-less GPU access (GLX-to-EGL interposer.)
We only ever really supported color index overlays anyhow, and color
index rendering was removed in OpenGL 3.1 (ten years ago.)  The feature
was originally developed for an oil & gas company whose users still had
3D workstations with Exceed 3D on their desks, so offloading the
rendering of overlays to the 2D X server wasn't a big deal.  These days,
however, the limitations of the feature are just too great to justify
the added complexity.