frameut: Fix occasional deadlock @ test completion
Bearing in mind that the producer/consumer relationship is as follows:

  Main thread (P) --> Compressor (C)
  Compressor (P) --> Decompressor (C)
  Decompressor (P) --> Blitter (C)

and that the threads are shut down from the consumer end first, a race
condition can occur whereby a producer calls the consumer's get() method
(to request the next frame from the consumer's pool) after the consumer
has been shut down, and since the get() method waits until processing is
complete on said frame, it will never return (since the consumer's
processing thread has been shut down, there is nothing to signal that
processing is complete.)

This commit modifies the get() methods such that they don't call
waitUntilComplete() if the consumer's processing thread is already dead.
If the processing thread is already dead, then there is no longer any
need to block the producer.  The producer may wastefully feed some new
data to the consumer that will never be processed, but the producer will
be shut down soon anyhow.