Faker: Fix DirectColor visual matching issues
It turns out that GLX_X_VISUAL_TYPE has no effect with
glXChooseFBConfig() if GLX_DRAWABLE_TYPE is specified but doesn't
contain GLX_WINDOW_BIT.  As such, the faker was not actually obtaining
DirectColor GLXFBConfigs to emulate the functionality of DirectColor
OpenGL visuals.  It doesn't technically need to, since those
GLXFBConfigs are only used for off-screen rendering, but it's easier and
more robust to store the visual type property with the GLXFBConfig
rather than the matching visual.  This commit does the following:

- Ensures that, in all internal calls to glXChooseFBConfig() that the
  faker makes in order to obtain GLXFBConfigs for use with off-screen
  rendering, GLX_WINDOW_BIT is added to GLX_DRAWABLE_TYPE if
  GLX_X_VISUAL_TYPE is specified.
- Modifies fakerut to verify that a DirectColor visual is obtained from
  VirtualGL when requesting GLX_X_VISUAL_TYPE=GLX_DIRECT_COLOR in
  glXChooseVisual() or glXChooseFBConfig() + glXGetVisualFromFBConfig().