Faker: Fix glXQueryContext[InfoEXT]()
The 3D X server screen number with which the context is actually
associated won't necessarily be the same as the 2D X server screen
number with which the context is supposed to be associated.  Also, when
glXQueryContextInfoEXT(GLX_VISUAL_ID_EXT) is called, we need to return
the 2D X server visual ID, not the 3D X server visual ID.  Go, go,
Gadget Context Hash.