Faker: Un-implement GLX_NV_swap_group
According to the extension spec [1], "GLX drawables except windows are
always ready to swap."  Thus, this extension doesn't currently do
anything meaningful in a VirtualGL environment.  It would need to be
emulated, and that emulation would be much more complicated than the
emulation of GLX_SGI_swap_control and GLX_EXT_swap_control.  Such
emulation would also be necessary in order to support GLX_NV_swap_group
with the upcoming EGL back end.  Given the dearth of applications that
use this extension, it hardly seems worth the trouble.

[1]
https://www.khronos.org/registry/OpenGL/extensions/NV/GLX_NV_swap_group.txt