Faker: Consistify WindowHash,PixmapHash interfaces
The bool return value for WindowHash.find() was a relic of transparent
overlay support.  This commit simplifies the interface such that
WindowHash.find() now has a single method signature, and for the sake of
readability, NULL is passed as the Display * argument in order to search
for a VirtualWin instance corresponding to a particular off-screen
drawable.