Remove #ifdefs for OpenGL 1.5
This was another relic of supporting antique platforms "back in the
day."  All of the platforms that VirtualGL currently supports have
OpenGL > 1.5.  Furthermore, PBO readback is no longer an optional
feature, and there are likely other parts of the VirtualGL code base
that would fail to compile on a system with OpenGL < 1.5.