vglrun: Prevent faker mismatch if LD_PRELOAD set
If a window manager is launched using vglrun, then LD_PRELOAD will be
set for every process launched within that window manager.  Because
vglrun appends the VGL fakers to LD_PRELOAD, launching an application
using vglrun with different -nodl or +ocl arguments than those used when
launching the window manager caused multiple flavors of the VGL Faker to
be loaded into the application.  VirtualGL (and, by extension, the
application) subsequently crashed due to one flavor of the faker loading
"real" functions from the other flavor of the faker.  This commit
modifies vglrun so that it removes any existing VGL fakers from
LD_PRELOAD before setting that environment variable, thus allowing
different -nodl and +ocl options to be used for a window manager and
individual applications launched within it.