glXDestroy*Pixmap(): Read back specified GLXPixmap
... rather than the current GLX drawable.  In the beginning, VirtualGL
always read back the current GLX drawable.  I don't remember why, but it
was probably incorrect behavior meant to work around other incorrect
behavior.  The scope of that behavior was restricted in
3a32a97c49b31260a6b56b236fa8baa4726f1f20 (VGL 2.4) to exclude
glXSwapBuffers(), since that function need not be called with the
current drawable.  However, that left the behavior in place for
glXDestroy[GLX]Pixmap(), which is also incorrect.  These are the only
readback triggers for which the underlying GLX drawable associated with
the VirtualPixmap or VirtualWin instance can differ from the current GLX
drawable, so reading back the underlying GLX drawable is always the
correct behavior.