Extend temp OGL ctx to cover all plugin functions
It may be necessary for some image transport plugins to create GPU-based
buffers or to otherwise use OpenGL within the body of RRTransGetFrame()
or RRTransInit().  Extending the temporary OpenGL context to cover all
image transport plugin functions comes with no cost and significantly
increases the flexibility of the plugin interface.