glXMakeContextCurrent(): Fix race condition
This was discovered while developing the EGL back end for VirtualGL 3.0.
If both read and draw GLXDrawables represented the same X window and
another thread resized that X window between the time that VGL's
interposed glXMakeContextCurrent() function processed the draw and read
GLXDrawables, then the underlying GLXPbuffer would be updated for the
draw GLXDrawable but not for the read GLXDrawable.  This led to
GLXBadDrawable or TempContext errors.  fakerut has been updated to test
for this issue.