FBXFrame/XVFrame: Mutex all calls to fbx[v]_init()
To back up and explain the situation, fbx[v]_init() relies on
XLockDisplay() in order to guard the global variables that it uses to
determine whether the MIT-SHM X extension is available.  However, since
the FBXFrame and XVFrame classes use a separate X display connection for
each instance, XLockDisplay() is insufficient to ensure the thread
safety of those classes.  99fdd4c163bc5a019889291ee5f207aae0a19a98
attempted to work around the problem by making the aforementioned global
variables thread-local, but XOpenDisplay() can still get into a data
race with fbx[v]_init() because it checks for the existence of an error
handler without first locking the Xlib global mutex.  The easiest
solution is just to revert the FBX fixes from
99fdd4c163bc5a019889291ee5f207aae0a19a98 and mutex the calls to
fbx[v]_init() and XOpenDisplay() in the FBXFrame and XVFrame classes.