EGL back end: Don't support glXCopyContext()
It proved impossible to implement that function in a robust manner, due
to the lack of a 1:1 correspondence between get and set functions for
certain OpenGL attributes and the fact that many of the attributes that
glXCopyContext() once handled are now obsolete.  glXCopyContext()
doesn't have broad driver support anyhow.  The implementation in AMD
Catalyst throws a BadRequest error, the implementation in VMWare throws
a BadMatch error, and the implementation in AMDGPU doesn't appear to
copy all of the attributes.  nVidia's implementation mostly works but
places certain attributes in the wrong category.  Thus, 3D applications
would be ill-advised to use glXCopyContext(), irrespective of VirtualGL.
glXCopyContext() was basically a companion for glPushAttrib() and
glPopAttrib(), which were removed in OpenGL 3.1 because they mostly
dealt with fixed-function state that was also removed.

This commit modifies the interposed version of glXCopyContext() so that
it throws BadRequest if the EGL back end is activated (thus emulating
the AMD Catalyst behavior.)  The change log has also been modified to
document which features are not supported or not yet supported by the
EGL back end.  (Support for GLX_texture_from_pixmap will be added prior
to release, if possible.)