EGL back end: Fix XCB; require 2D X server GLX ext
If the 2D X server has no GLX extension, then there is no sane way to
emulate XCB-GLX with the EGL back end.  Our hackish attempt to handle
GLX errors for Xlib programs, in the absense of a GLX extension, was
problematic, because we were attempting to create a fictional X
extension that wasn't registered with the 2D X server.  Thus, the GLX
error strings weren't correct (causing fakerut to fail.)  That problem
could have been fixed by interposing XGetErrorText(), but there was
still no guarantee that we wouldn't stomp on a legitimate 2D X server
extension.  And it doesn't make much sense to support the EGL back end
with Xlib and not XCB.  As much as it would be nice to support the EGL
back end with 2D X servers that lack a GLX extension, the reality is
that most, if not all, such 2D X servers are obsolete.  TigerVNC has
always had a GLX extension, and it shouldn't be a big deal to require
TurboVNC 2.2+ for a next-gen VirtualGL feature.