EGL back end: Eliminate incorrect uses of DPY3D
In faker-glx.cpp, DPY3D was incorrectly passed to VGLSwapBuffers() and
VGLQueryContext() rather than dpy.  This caused a segfault if either
function tried to throw a GLX error.

Guard the uses of DPY3D in the interposed glXCreate*Context*()
functions.  This is really a belt-and-suspenders measure, since the EGL
back end can never create an indirect context.

Change references to (EGLDisplay)DPY3D to EDPY, for readability (they're
really the same thing.)