EGLBE: Don't pass NULL ctx to eglDestroyContext()
For reasons unexplained, nVidia's implementation of eglDestroyContext()
returns EGL_FALSE but doesn't set the EGL error to EGL_BAD_CONTEXT if
passed a NULL context.  The AMD Catalyst, nVidia, and VMWare
implementations of glXDestroyContext() do not throw a GLXBadContext
error if passed a NULL context, so emulate that behavior.  This issue
was discovered while trying to run the Qt5 computegles31 example.