Faker: Don't trace trivial OpenGL wrappers
VGL_TRACE is primarily intended as a diagnostic aid for
incompatibilities between 3D applications and the VirtualGL front end
(the GLX interposer and the Window-to-Pbuffer emulation layer.)  Thus,
it doesn't make sense to trace non-GLX functions that are trivial
wrappers with the GLX back end, i.e. non-GLX functions that have no
front end code and are interposed solely to support the EGL back end.