glXChooseVisual: Don't choose FB config w/o visual
Certain GLX implementations, such as AMDGPU and Nouveau, put 10-bpc
GLXFBConfigs at the head of the list, and these 10-bpc GLXFBConfigs have
GLX_PIXMAP_BIT set but no X visual attached (which is arguably
incorrect.)  VirtualGL's implementation of glXChooseVisual() chose the
first GLXFBConfig returned by the "real" glXChooseFBConfig() function,
so when used with the aforementioned GLX implementations, a 10-bpc
GLXFBConfig was erroneously chosen behind the scenes if the 3D
application requested an 8-bpc visual.  This caused the VGL Transport to
fail in environments where it would normally work.