Faker: Fix img flip w/ X11Trans/glFlush/!GL_RENDER
Previously, if the render mode was not GL_RENDER,
VirtualDrawable::readPixels() would refuse to read back the framebuffer,
but VirtualWin::readback() would still attempt to transport a frame.
This caused a previously-rendered frame to be re-transported.  When
using the X11 Transport, the bottom-up images read back from the
framebuffer are flipped vertically prior to display, since
X[Shm]PutImage() expects top-down images.  Thus, when the
previously-rendered frame was re-transported, the FBXFrame instance
already contained a top-down image, and re-transporting it caused
fbx_flip() to be called unnecessarily, converting the image back to
bottom-up.

This issue was known to affect Altair HyperWorks 2019, and it probably
affected prior versions as well.