vglserver_config: Fix issues if invoked w/rel path
Previously, if a user executed (for example)

    cd /opt/VirtualGL/bin
    sudo ./vglserver_config

then vglserver_config would add "./vglgenkey" to the display manager
startup scripts, which is incorrect.  Using realpath to determine the
absolute path of vglserver_config fixes this issue, and it has the added
benefit of following symlinks.  Thus, invoking

    sudo /usr/bin/vglserver_config
or
    sudo /bin/vglserver_config

with our official packages will cause /opt/VirtualGL/bin/vglgenkey to be
added to the display manager startup scripts, and the configured system
will not rely upon the presence of the symlinks.

Fixes #147