Faker: Work around AMDGPU issue w/ Google Chrome
The AMDGPU GLX implementation apparently lacks glXSwapIntervalEXT(),
which Chrome needs.  VGL's interposed version of glXGetProcAddress[ARB]()
was unnecessarily strict, checking for the presence of
glXSwapIntervalEXT() in the underlying GLX implementation even though
the interposed version of glXSwapIntervalEXT() never calls the
underlying implementation unless overlay rendering is enabled or the
X display is excluded.  Because the interposed version of
glXGetProcAddress[ARB]() returned NULL for glXSwapIntervalEXT(), Chrome
fell back to using EGL/X11, which we don't support yet (refer to #66),
and failing to intialize EGL, it disabled GPU acceleration.

NOTE: This commit also pre-emptively fixes the same issue with
glXSwapIntervalSGI().

Fixes #146