Faker: Interpose XkbOpenDisplay()
XkbOpenDisplay() calls XOpenDisplay(), but since that function call
occurs within libX11, VirtualGL cannot intercept it on some platforms.
(Specifically, VGL picks up the internal XOpenDisplay() call on CentOS
but not Ubuntu.)  Thus we need to interpose XkbOpenDisplay().

The downside for not doing this in VGL 2.6.x is minimal and not worthy
of documentation, since it is not known to affect any real-world
applications:

- ServerVendor(), if invoked with the display handle returned from
  XkbOpenDisplay(), does not return the value of VGL_XVENDOR (but
  XServerVendor() does.)  VGL_XVENDOR was implemented to support
  Pro/ENGINEER v2.0 on Solaris/SPARC in 2006.  VGL_XVENDOR is not known
  to be necessary with any modern applications, and if memory serves,
  even Pro/E v2.0 used XServerVendor() rather than ServerVendor().

- VGL_EXCLUDE is not honored for any display connection opened using
  XkbOpenDisplay() rather than XOpenDisplay().

The downside for not doing this in VGL 3.0, however, is more severe.
Referring to #153, at least some applications that use XkbOpenDisplay()
rather than XOpenDisplay() stopped working properly when
0168619c41d5b70928ad220b34d3fd7066b20798 was introduced.