FBX:Disable NoExpose X events if not using MIT-SHM
Otherwise, if the MIT-SHM X extension is unavailable because:

- the 2D X server does not support MIT-SHM,
- MIT-SHM is explicitly disabled in the 2D X server (e.g. by running
  '/opt/TurboVNC/bin/vncserver -extension MIT-SHM'),
- MIT-SHM is explicitly disabled in VirtualGL (e.g. by setting
  VGL_USEXSHM=0 or running 'fbxtest -noshm'), or
- the application does not share an IPC namespace with the 2D X server
  (which is the case with remote X connections)

and the 2D X server is using a "traditional" window manager (e.g. TWM)
or no window manager, X NoExpose events will queue up in Xlib until
the 3D application exits or the memory on the VirtualGL server is
exhausted.

Based on:
https://github.com/tghosgor/virtualgl/commit/30a68c3d14c2a1f46e3a8d93cfb08959dec3c688

Fixes #155
Closes #156