EGL: Work around segv w/cgroup-blacklisted NV devs
Referring to #157, this commit works around an issue in v418.xx-450.xx
(inclusive) of nVidia's proprietary OpenGL stack whereby
eglQueryDevicesEXT() returns all devices, including devices that are
blacklisted using control groups (cgroups), but
eglQueryDeviceStringEXT() subsequently segfaults when passed a
cgroup-blacklisted device.  Both the VirtualGL Faker and eglinfo now
attempt to call eglGetPlatformDisplayEXT() and eglInitialize() for every
device returned from eglQueryDevicesEXT(), skipping any devices for
which either of those functions fails (eglInitialize() will fail when
passed a cgroup-blacklisted device.)

This allows cgroups to be used with the EGL back end and v418.xx-450.xx
of the nVidia proprietary drivers, although cgroups unfortunately cannot
be used with v460.xx (eglQueryDevicesEXT() in 460.xx returns no devices
if any of the devices is cgroup-blacklisted.)

This commit also modifies eglinfo such that it allows a device name of
"egl" to be used to specify the first available EGL device, as the
VirtualGL Faker does.