Faker: Work around bugs in the AMDGPU driver
This commit causes the VirtualGL Faker to work around a couple of
GLX conformance issues in the AMDGPU driver if an undocumented
environment variable (VGL_AMDGPUHACK) is set to 1.  AMD was notified
about these conformance issues almost a year ago, so it is hoped that
they will eventually be fixed.  This hack will be removed once that is
the case.

GLX conformance issues
----------------------

- glXSwapBuffers() does not actually swap the buffers of a
  double-buffered Pbuffer.

  Reference:
  https://www.khronos.org/registry/OpenGL-Refpages/gl2.1/xhtml/glXCreatePbuffer.xml
  "GLXPbuffers contain the color and ancillary buffers specified by
  config.  It is possible to create a pixel buffer with back buffers and
  to swap those buffers using glXSwapBuffers."

- glXQueryContext(..., GLX_RENDER_TYPE, ...) returns a render type of 0
  if the context was created with a render type of GLX_RGBA_TYPE.

  Reference:
  https://www.khronos.org/registry/OpenGL-Refpages/gl2.1/xhtml/glXQueryContext.xml
  "GLX_RENDER_TYPE: Returns the rendering type supported by ctx."