EGLBE: Fix multisamp glReadPixels buf swap w/ Mesa
Referring to the previous commit message, Mesa seems to automatically
set the draw/read buffer to GL_COLOR_ATTACHMENT0 within the body of
glBindFramebuffer() when binding an FBO, regardless of whether another
FBO was already bound.  Thus, we need to handle Mesa's behavior in
VGLReadPixels() when multisampling is enabled, since that function
temporarily creates and binds a new FBO container.